bastion.scripts/requirement.txt
boto3
requests
mysql-connector-python
